% Auto-generated: do not edit by hand
\docType{package}
\name{dashCytoscape-package}
\alias{dashCytoscape}
\title{Interactive Network Visualization Component for Dash
}
\description{
A Component Library for Dash aimed at facilitating network visualizat    ion in R, wrapped around Cytoscape.js.

}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly>
}
