% Auto-generated: do not edit by hand
\docType{package}
\name{dashCytoscape-package}
\alias{dashCytoscape}
\title{A Component Library for 'Dash' aimed at facilitating network visualization in R, wrapped around Cytoscape.js.}
\description{
A Component Library for 'Dash' aimed at facilitating network visualization in R, wrapped around Cytoscape.js.
}
\author{
\strong{Maintainer}: The Plotly Team <cytoscape@plotly.com>
}
